+++ tests/test_utils.py
@@ -1,13 +1,16 @@
+from collections import Counter
 from sys import maxsize
 
 import pytest
 from scooze.catalogs import Format
 from scooze.utils import (
+    CostSymbol,
     DictDiff,
     cmdr_size,
     main_size,
     max_card_quantity,
     max_relentless_quantity,
+    parse_symbols,
     side_size,
 )
 
@@ -140,6 +143,35 @@ def cmdr_size_2() -> tuple[int, int]:
 @pytest.fixture
 def cmdr_size_any() -> tuple[int, int]:
     return 0, maxsize
+
+
+# endregion
+
+
+# region Mana costs
+@pytest.fixture
+def mana_cost_one_symbol() -> str:
+    return "{W}"
+
+
+@pytest.fixture
+def mana_cost_multiple_symbols() -> str:
+    return "{W}{B}"
+
+
+@pytest.fixture
+def mana_cost_with_generic() -> str:
+    return "{2}{W}{B}"
+
+
+@pytest.fixture
+def mana_cost_with_multiple_generic() -> str:
+    return "{1}{1}{U}{R}"
+
+
+@pytest.fixture
+def mana_cost_with_hybrid() -> str:
+    return "{B/G}{B/G}"
 
 
 # endregion
@@ -736,6 +768,37 @@ def test_fmt_none_cmdr_size(cmdr_size_any):
 
 
 # endregion
+
+# endregion
+
+# region Mana symbology
+
+
+def test_parse_symbols_one_symbol(mana_cost_one_symbol):
+    assert parse_symbols(mana_cost_one_symbol) == {CostSymbol.WHITE: 1}
+
+
+def test_parse_symbols_multiple_symbols(mana_cost_multiple_symbols):
+    assert parse_symbols(mana_cost_multiple_symbols) == {CostSymbol.WHITE: 1, CostSymbol.BLACK: 1}
+
+
+def test_parse_symbols_with_generic(mana_cost_with_generic):
+    assert parse_symbols(mana_cost_with_generic) == {CostSymbol.WHITE: 1, CostSymbol.BLACK: 1, CostSymbol.GENERIC_2: 1}
+
+
+def test_parse_symbols_with_multiple_generic(mana_cost_with_multiple_generic):
+    assert parse_symbols(mana_cost_with_multiple_generic) == {
+        CostSymbol.BLUE: 1,
+        CostSymbol.RED: 1,
+        CostSymbol.GENERIC_1: 2,
+    }
+
+
+def test_parse_symbols_with_hybrid(mana_cost_with_hybrid):
+    assert parse_symbols(mana_cost_with_hybrid) == {
+        CostSymbol.HYBRID_BG: 2,
+    }
+
 
 # endregion